refactor: 그룹 당당자 추가, 조회 기능 파일 분리
